+++ examples/index.php
@@ -245,7 +245,19 @@ require_once __DIR__ . '/../../../autoload.php';
                 <div class="card olive">
                     <div class="content">
                         <div class="header">
-                            Flexipay direct (PIS)
+                            FlexiPay Rate Direct Debit (Hire Purchase)
+                        </div>
+                        <div class="description">
+                        </div>
+                    </div>
+                    <div class="ui bottom attached green button" onclick="location.href='HirePurchaseDirectDebit/';">
+                        Try
+                    </div>
+                </div>
+                <div class="card olive">
+                    <div class="content">
+                        <div class="header">
+                            FlexiPay Direct (PIS)
                         </div>
                         <div class="description">
                         </div>