+++ test/integration/TransactionTypes/ChargeTest.php
@@ -105,12 +105,12 @@ class ChargeTest extends BasePaymentTest
         $paymentReference = 'paymentReference';
 
         // perform request
-        $charge = $paymentType->charge(123.4, 'EUR', self::RETURN_URL, $customer, $orderId, $metadata, $basket, true, $invoiceId, $paymentReference);
+        $charge = $paymentType->charge(100.19, 'EUR', self::RETURN_URL, $customer, $orderId, $metadata, $basket, true, $invoiceId, $paymentReference);
 
         // verify the data sent and received match
         $payment = $charge->getPayment();
         $this->assertSame($paymentType, $payment->getPaymentType());
-        $this->assertEquals(123.4, $charge->getAmount());
+        $this->assertEquals(100.19, $charge->getAmount());
         $this->assertEquals('EUR', $charge->getCurrency());
         $this->assertEquals(self::RETURN_URL, $charge->getReturnUrl());
         $this->assertSame($customer, $payment->getCustomer());
@@ -155,12 +155,12 @@ class ChargeTest extends BasePaymentTest
         $paymentReference = 'paymentReference';
 
         // perform request
-        $charge = $ivg->charge(123.4, 'EUR', self::RETURN_URL, $customer, $orderId, $metadata, $basket, null, $invoiceId, $paymentReference);
+        $charge = $ivg->charge(100.19, 'EUR', self::RETURN_URL, $customer, $orderId, $metadata, $basket, null, $invoiceId, $paymentReference);
 
         // verify the data sent and received match
         $payment = $charge->getPayment();
         $this->assertSame($ivg, $payment->getPaymentType());
-        $this->assertEquals(123.4, $charge->getAmount());
+        $this->assertEquals(100.19, $charge->getAmount());
         $this->assertEquals('EUR', $charge->getCurrency());
         $this->assertEquals(self::RETURN_URL, $charge->getReturnUrl());
         $this->assertSame($customer, $payment->getCustomer());